test: refactor messages.json validity check into integration test
